% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosenbrock.R
\name{x1_iter}
\alias{x1_iter}
\title{The iterative procedure for the x1 variable, derived by minimizing the surrogate for x1
see Lange for details of the inequality used to derive the surrogate}
\usage{
x1_iter(x1, x2)
}
\arguments{
\item{x1}{scalar}

\item{x2}{scalar}
}
\value{
next iterate of the x1 variable
}
\description{
The iterative procedure for the x1 variable, derived by minimizing the surrogate for x1
see Lange for details of the inequality used to derive the surrogate
}
\examples{
x1_iter(1,1)
}
