% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosenbrock.R
\name{x2_iter}
\alias{x2_iter}
\title{Iterative procedure for the x2 variable, derived by minimizing the surrogate for x2
see Lange for details of the inequality used to derive the surrogate}
\usage{
x2_iter(x1, x2)
}
\arguments{
\item{x1}{scalar}

\item{x2}{scalar}
}
\value{
next iterate of the x2 variable
}
\description{
Iterative procedure for the x2 variable, derived by minimizing the surrogate for x2
see Lange for details of the inequality used to derive the surrogate
}
\examples{
x2_iter(1,1)
}
